+++ js/data.js
@@ -1,0 +1,33 @@
+const DESCRIPTION_COUNT = 25;
+const LIKE_MIN = 15;
+const LIKE_MAX = 200;
+const COMMENTS_MAX = 30;
+const AVATAR_ID_MAX = 6;
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const NAMES = [
+  'Иван',
+  'Артем',
+  'Андрей',
+  'Георгий',
+  'Василий'
+];
+
+const DESCRIPTIONS = [
+  'Центр города',
+  'Окраина',
+  'Парк',
+  'Каньон',
+  'Деревня',
+  'Площадка'
+];
+
+export {MESSAGES, NAMES, DESCRIPTIONS, DESCRIPTION_COUNT, LIKE_MIN, LIKE_MAX, COMMENTS_MAX, AVATAR_ID_MAX};
